requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
faiss-cpu==1.11.0
filelock==3.13.1
Flask==3.1.1
fonttools==4.58.1
fsspec==2024.6.1
ftfy==6.3.1
idna==3.10
importlib-metadata==8.7.0
importlib-resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.3
numpy==2.2.6
packaging==25.0
Pillow==11.2.1
pyparsing==3.2.3
python-dateutil==2.9.0
regex==2024.11.6
requests==2.32.3
six==1.17.0
sympy==1.13.3
torch==2.7.1+cu126
torchaudio==2.7.1+cu126
torchvision==0.22.1+cu126
tqdm==4.67.1
typing-extensions==4.14.0
urllib3==2.4.0
wcwidth==0.2.13
Werkzeug==3.1.3
zipp==3.22.0
MarkupSafe==2.1.3
matplotlib==3.7.2
numpy==1.24.1
packaging==23.1
Pillow==9.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
regex==2023.6.3
requests==2.28.1
six==1.16.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.65.0
typing_extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.6
Werkzeug==2.3.6
zipp==3.16.2
